Use extensions for application output and statistics
My goal is to make the application class simpler. It should not know whether
it's being run as a CLI tool nor know anything about outputting results.

Currently the Application still keeps track and shows which files are being
tested. This will be changed once instead of using simple strings for events we
use objects that contain associated data.