Emit event objects instead of strings
Now that we emit objects instead of plain old strings we can give the objects
data such as the mutation that has been applied.

The extensions have been changed to use these objects and we've added a
KeepTrackOfMutationsOnFile extension that shows the amount of mutations that
have been applied on a file.